add payment_data
