+++ lib/hotel_beds/model/payment_data.rb
@@ -1,0 +1,12 @@
+require "hotel_beds/model"
+
+module HotelBeds
+  module Model
+    class PaymentData
+      include HotelBeds::Model
+
+      # attributes
+      attribute :description, String
+    end
+  end
+end
