+++ lib/hotel_beds/parser/reference.rb
@@ -2,15 +2,11 @@ require "hotel_beds/parser"
 
 module HotelBeds
   module Parser
-    class Reference
+    class PaymentData
       include HotelBeds::Parser
 
       # attributes
-      attribute :file_number, selector: "FileNumber"
-      attribute :incoming_office_code, {
-        selector: "IncomingOffice",
-        attr: "code"
-      }
+      attribute :description, selector: "Description"
     end
   end
 end
